Rafactor questions & answers with concerns
