+++ app/assets/javascripts/votes.coffee
@@ -5,15 +5,15 @@
 $(document).on 'turbolinks:load', ->
   update_rating = (object) ->
     object.bind 'ajax:success', (e, data, status, xhr) ->
-      rating = $.parseJSON(xhr.responseText)
-      $('div.current_rating').html(rating)
+      votable = $.parseJSON(xhr.responseText)
+      $('div.current_rating').html(votable.rating)
     .bind 'ajax:error', (e, xhr, data, error) ->
       $('div.rating_errors').html(xhr.responseText)
 
-  $(document).on 'click', 'a.vote-up-question, a.vote-down-question', (e) ->
+  $(document).on 'click', 'a.vote-up, a.vote-down', (e) ->
     update_rating($(this))
-    $('div.rating a.revote-question').show()
+    $('div.rating a.revote').show()
 
-  $(document).on 'click', 'a.revote-question', (e) ->
+  $(document).on 'click', 'a.revote', (e) ->
     update_rating($(this))
     $(this).hide()
