+++ app/controllers/questions_controller.rb
@@ -1,6 +1,7 @@
 class QuestionsController < ApplicationController
-  before_action :authenticate_user!, except: [:index, :show]
-  before_action :set_question, only: [:show, :destroy, :update, :vote_up, :vote_down, :revote]
+  include Voted
+
+  before_action :set_question, only: [:show, :destroy, :update]
 
   def index
     @questions = Question.all
@@ -43,34 +44,7 @@ class QuestionsController < ApplicationController
     end
   end
 
-  def vote_up
-    vote('vote_up')
-  end
-
-  def vote_down
-    vote('vote_down')
-  end
-
-  def revote
-    if @question.voted?(current_user)
-      @question.revote(current_user)
-
-      respond_to { |format| format.json { render json: @question.votes.count_rating } }
-    else
-      render :show
-    end
-  end
-
   private
-
-  def vote(choice)
-    if @question.voted?(current_user) || current_user.author_of?(@question)
-      respond_to { |format| format.json { render json: t('vote_error'), status: :unprocessable_entity } }
-    else
-      @question.send(choice, current_user)
-      respond_to { |format| format.json { render json: @question.votes.count_rating } }
-    end
-  end
 
   def set_question
     @question = Question.find(params[:id])
