+++ app/models/answer.rb
@@ -1,10 +1,9 @@
 class Answer < ApplicationRecord
+  include Attachable
+  include Votable
+
   belongs_to :question
   belongs_to :user
-  has_many :attachments, as: :attachable
-  has_many :votes, as: :votable, dependent: :destroy
-
-  accepts_nested_attributes_for :attachments, reject_if: :all_blank, allow_destroy: true
 
   validates :body, presence: true
