+++ app/views/common/_rating.html.slim
@@ -1,0 +1,16 @@
+.rating
+  - if resource.votes.any?
+    .current_rating
+      = resource.votes.count_rating
+      .rating_errors
+  - else
+    p 0
+  - unless current_user&.author_of?(resource)
+    p
+      => link_to '+1', { controller: resource.to_string, id: resource, action: 'vote_up' }, class: 'vote-up', method: :patch, remote: true, data: { type: :json, "#{resource.to_string}": resource.id }
+      |
+        /
+      =< link_to '-1', { controller: resource.to_string, id: resource, action: 'vote_down' }, class: 'vote-down', method: :patch, remote: true, data: { type: :json }
+    - if user_signed_in? && resource.voted?(current_user)
+      .revote
+        p= link_to 'Revote', { controller: resource.to_string, id: resource, action: 'revote' }, class: 'revote', method: :patch, remote: true, data: { type: :json }
