+++ config/routes.rb
@@ -8,7 +8,7 @@ Rails.application.routes.draw do
   end
 
   resources :questions, concerns: :votable do
-    resources :answers, shallow: true do
+    resources :answers, concerns: :votable, shallow: true do
       patch :best, on: :member
     end
   end
